# microservice-notification/swagger.yaml
{"openapi":"3.0.1","info":{"title":"Microservice Notification","description":"Сервис оповещений"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Notification service","description":"Сервис оповещений"}],"paths":{"/api/v1/notifications/settings":{"get":{"tags":["Notification service"],"description":"Получение настроек оповещений","operationId":"getSetting","responses":{"401":{"description":"Unauthorized"},"400":{"description":"Bad request"},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"$ref":"#/components/schemas/NotificationSettingDto"}}}}},"security":[{"JWT":[]}]},"put":{"tags":["Notification service"],"description":"Коррекция настроек оповещений","operationId":"updateSetting","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotificationUpdateDto"}}},"required":true},"responses":{"401":{"description":"Unauthorized"},"400":{"description":"Bad request"},"200":{"description":"Successful operation"}},"security":[{"JWT":[]}]}},"/api/v1/notifications/readed":{"put":{"tags":["Notification service"],"description":"Отметить все события, как прочитанные","operationId":"setIsRead","responses":{"401":{"description":"Unauthorized"},"400":{"description":"Bad request"},"200":{"description":"Successful operation"}},"security":[{"JWT":[]}]}},"/api/v1/notifications/settings{id}":{"post":{"tags":["Notification service"],"description":"Создание настроек оповещений","operationId":"createSetting","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"401":{"description":"Unauthorized"},"400":{"description":"Bad request"},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"boolean"}}}}},"security":[{"JWT":[]}]}},"/api/v1/notifications/add":{"post":{"tags":["Notification service"],"description":"Создание события","operationId":"addEvent","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventNotificationDto"}}},"required":true},"responses":{"401":{"description":"Unauthorized"},"400":{"description":"Bad request"},"200":{"description":"Successful operation"}},"security":[{"JWT":[]}]}},"/api/v1/notifications":{"get":{"tags":["Notification service"],"description":"Получение событий","operationId":"getNotifications","parameters":[{"name":"page","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"401":{"description":"Unauthorized"},"400":{"description":"Bad request"},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageNotificationsDto"}}}}},"security":[{"JWT":[]}]}},"/api/v1/notifications/count":{"get":{"tags":["Notification service"],"description":"Получить счетчик количества событий","operationId":"getCount","responses":{"401":{"description":"Unauthorized"},"400":{"description":"Bad request"},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"$ref":"#/components/schemas/NotificationCountDto"}}}}},"security":[{"JWT":[]}]}}},"components":{"schemas":{"NotificationUpdateDto":{"type":"object","properties":{"enable":{"type":"boolean","description":"Разрешить оповещение для данного типа событий"},"notificationType":{"type":"string","description":"Тип события"}},"description":"Dto для установки настроек оповещений"},"EventNotificationDto":{"type":"object","properties":{"authorId":{"type":"string","format":"uuid"},"receiverId":{"type":"string","format":"uuid"},"notificationType":{"type":"string"},"content":{"type":"string"}}},"Pageable":{"type":"object","properties":{"page":{"minimum":0,"type":"integer","format":"int32"},"size":{"minimum":1,"type":"integer","format":"int32"},"sort":{"type":"array","items":{"type":"string"}}}},"NotificationDto":{"type":"object","properties":{"id":{"type":"string","description":"Id","format":"uuid"},"authorId":{"type":"string","description":"Id пользователя, создавшего событие","format":"uuid"},"content":{"type":"string","description":"Описание события"},"notificationType":{"type":"string","description":"Тип события"},"sentTime":{"type":"string","description":"Время события","format":"date-time"}},"description":"Dto события доп."},"NotificationsDto":{"type":"object","properties":{"timeStamp":{"type":"string","description":"Врямя отправки события","format":"date-time"},"data":{"$ref":"#/components/schemas/NotificationDto"}},"description":"Dto события"},"PageNotificationsDto":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"number":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/NotificationsDto"}},"sort":{"$ref":"#/components/schemas/Sort"},"first":{"type":"boolean"},"last":{"type":"boolean"},"numberOfElements":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"offset":{"type":"integer","format":"int64"},"sort":{"$ref":"#/components/schemas/Sort"},"paged":{"type":"boolean"},"pageSize":{"type":"integer","format":"int32"},"unpaged":{"type":"boolean"},"pageNumber":{"type":"integer","format":"int32"}}},"Sort":{"type":"object","properties":{"empty":{"type":"boolean"},"unsorted":{"type":"boolean"},"sorted":{"type":"boolean"}}},"NotificationSettingDto":{"type":"object","properties":{"id":{"type":"string","description":"Id","format":"uuid"},"enableLike":{"type":"boolean","description":"Разрешить события с типом LIKE"},"enablePost":{"type":"boolean","description":"Разрешить события с типом POST"},"enablePostComment":{"type":"boolean","description":"Разрешить события с типом POST_COMMENT"},"enableCommentComment":{"type":"boolean","description":"Разрешить события с типом COMMENT_COMMENT"},"enableMessage":{"type":"boolean","description":"Разрешить события с типом MESSAGE"},"enableFriendRequest":{"type":"boolean","description":"Разрешить события с типом FRIEND_REQUEST"},"enableFriendBirthday":{"type":"boolean","description":"Разрешить события с типом FRIEND_BIRTHDAY"},"enableSendEmailMessage":{"type":"boolean","description":"Разрешить события с типом SEND_EMAIL_MESSAGE"}},"description":"Dto настроек оповещений"},"Count":{"type":"object","properties":{"count":{"type":"integer","description":"Значение счетчика событий","format":"int64"}},"description":"Dto получения счетчика событий доп."},"NotificationCountDto":{"type":"object","properties":{"timeStamp":{"type":"string","description":"Время выдачи показаний счетчика","format":"date-time"},"data":{"$ref":"#/components/schemas/Count"}},"description":"Dto получения счетчика событий"}},"securitySchemes":{"JWT":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}